Switching to Go Modules
As of Go 1.11 it is possible to use Go Modules. This change adds the Go
modules file, the auto generated go checksum file and also the generated
vendor folder.

The imort path for logrus had to be changed, because sirupsen changed his
account name to be fully lower case and go mod was choking on the old
upper case version.