Adds readiness and liveness notifications for systemd
Once the server is started a readiness notification is send to systemd. If
the service is using "Type=notify", we can tell systemd that Filtra is
ready.

A liveliness notification is send periodically after checking if port 8080
is answering. This is then used by the systemd watchdog to check if the
service should be restarted.